+++ chpw-cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/keith-hung/chpw-cli/cmd"
+
+func main() {
+	cmd.Execute()
+}